requirements.txt
fastapi
uvicorn
rapidfuzz
requests
python-dotenv
pydantic
python-dotenv